+++ examples/async/client_callchain.py
@@ -11,6 +11,11 @@ def asyncFunction(value, increase=1):
     print(">>> async function called with value={0} increase={1}".format(value,increase))
     return value+increase
 
+def asyncWithMoreArgs(a, b, extra=None):
+    print(">>> async func called with some arguments: a={0}, b={1}, extra={2}".format(a, b, extra))
+    return a+b
+
+
 uri=input("enter async server object uri: ").strip()
 proxy=Pyro4.Proxy(uri)
 
@@ -43,18 +48,19 @@ assert asyncresult.value==23
 
 print("\n* async call with call chain, where calls have extra arguments:")
 async=Pyro4.async(proxy)
-asyncresult=async.divide(100,5)
+asyncresult=async.multiply(5,6)
 # set a chain of callables to be invoked once the value is available
-# the callable will be invoked like so:  function(asyncvalue, kwarg=..., kwarg=...)
-asyncresult.then(asyncFunction, increase=1) \
-           .then(asyncFunction, increase=2) \
-           .then(asyncFunction, increase=3)
-print("sleeping 4 seconds")
-time.sleep(4)   # the call chain will be invoked during this sleep period
+# the callable will be invoked like so:  function(asyncvalue, normalarg, kwarg=..., kwarg=...)
+# (the value from the previous call is passed as the first argument to the next call)
+asyncresult.then(asyncWithMoreArgs, 1) \
+           .then(asyncWithMoreArgs, 2, extra=42) \
+           .then(asyncWithMoreArgs, 3, extra="last one")
+print("sleeping 1 second")
+time.sleep(1)   # the call chain will be invoked during this sleep period
 print("back from sleep")
-# you can still access the final asyncresult.value. It should be 100/5+1+2+3=26
+# you can still access the final asyncresult.value. It should be 5*6+1+2+3=36
 print("final value=",asyncresult.value)
-assert asyncresult.value==26
+assert asyncresult.value==36
 
 print("\n* async call with call chain, where calls are new pyro calls:")
 async=Pyro4.async(proxy)