+++ docs/source/changelog.rst
@@ -4,8 +4,12 @@ Change Log
 
 **Pyro 4.10**
 
+- added Future class that provides async (future) function calls for any callable (not just Pyro proxies)
+- renamed _AsyncResult to FutureResult
+- improved FutureResult.then(), it now accepts additional normal arguments as well instead of only kwargs
 - renamed Pyro4.config.refresh to Pyro4.config.reset because reset better describes what it is doing
 - added parameter to config.refresh to make it ignore environment variables
+
 
 **Pyro 4.9**